A simple stackmap parser.
Currently only the information that we will obviously need for yorickrt
is being exposed, but we can extract additional info in a later change
if required.

The test suite currently requires LLVM-6.0 with binaries suffixed
`-6.0`. LLVM bytecodes containing stackmap call are compiled as test
inputs, then the results of `llvm-readelf -stackmap` are compared to the
results of the library.