Refactor llvm-readobj test parser to be extensible
Parsing the output of llvm-readobj is fiddly and it's not really
possible to treat each line uniformly. Currently, we only parse a
stackmap record's patchpoint ID and instruction offset, but we will want
to be able parse everything. Adding this functionality to the same
stdout line iterator will quickly become unmaintainable.

This splits the parsing of each main stackmap 'block' into distinct
functions: a function parser, and a record parser (with more in future
when we need to parse additional output)