Use patched version of llvm-readobj for testing
The current LLVM StackMap version is 3. This is *not* backwards
compatible with the previous versions. Unfortunately, the current
version of llvm-readobj provided with llvm does not support parsing
version 3 stackmaps. I've submitted a patch to llvm [0] but it will be a
while until this makes it upstream.

In the meantime, we can use the environment variable "LLVM_READOBJ_PATH"
to tell cargo the path of our patched readobj bin. There are a couple of
options for users to download and get the modified readobj binary
themselves:

 1. Build from source using my fork of llvm on github [1]
 2. Download a prebuilt bin from this archive repo [2]

 Travis will fetch a prebuilt binary from the archive repo in [2] and
 use it during the build.

 [0] https://reviews.llvm.org/D59020
 [1] https://github.com/jacob-hughes/llvm-project/commit/796c0167def7de136e36e93479eb57c571a523c2
 [2] https://github.com/jacob-hughes/llvm-readobj-bin