Use patched version of llvm-readobj for testing
The current LLVM StackMap version is 3. This is *not* backwards
compatible with the previous versions. Unfortunately, the current
version of llvm-readobj provided with llvm does not support parsing
version 3 stackmaps. I've submitted a patch to llvm [0] but it will be a
while until this makes it upstream.

In the meantime, we can use the environment variable "LLVM_READOBJ_PATH"
to tell cargo the path of our patched readobj bin. Users can download a
modified version of llvm-readobj from my fork of LLVM [1].

 [0] https://reviews.llvm.org/D59020
 [1] https://github.com/jacob-hughes/llvm-project/commit/796c0167def7de136e36e93479eb57c571a523c2