Parse the location block in stackmap record entries
Locations give vital information about what values have been spilled to
the stack at a given call site. This is useful in GC for identifying:
which values spilled to the stack are root pointers; and where, at the
given callsite, they exist in the runtime (register, on-stack, as a
constant etc).

The location block for version 3 stackmaps has the following format*:

  Location[NumLocations] {
    uint8  : Register | Direct | Indirect | Constant | ConstantIndex
    uint8  : Reserved (expected to be 0)
    uint16 : Location Size
    uint16 : Dwarf RegNum
    uint16 : Reserved (expected to be 0)
    int32 | uint32  : Offset or SmallConstant
  }

* The type of the Offset or SmallConstant field (int32 | uint32) is
specified in the documentation as an in32. However, this differs in the
implementation. I have chosen to use LLVMs implementation details
here as the source of truth (See comment src/lib.rs:86-100)